pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-proxy-bot"
version = "0.1.0"
description = "Telegram bot for distributing private MTProto proxy access by subscription"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.6,<4",
  "sqlalchemy>=2.0,<3",
  "psycopg[binary]>=3.1,<4",
  "python-dotenv>=1.0,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]
